+++ tests/testthat/test-joins.r
@@ -370,3 +370,24 @@ test_that("join columns are not moved to the left (#802)", {
   out <- left_join(df1, df2)
   expect_equal(names(out), c("x", "y", "z"))  
 })
+
+test_that("join can handle multiple encodings (#769)", {
+  x <- data_frame(name=c("\xC9lise","Pierre","Fran\xE7ois"),score=c(5,7,6))
+  y <- data_frame(name=c("\xC9lise","Pierre","Fran\xE7ois"),attendance=c(8,10,9))
+  res <- left_join(x, y, by = "name")
+  expect_equal( nrow(res), 3L)
+  expect_equal( res$x, x$x)
+  
+  x <- data_frame(name=factor(c("\xC9lise","Pierre","Fran\xE7ois")),score=c(5,7,6))
+  y <- data_frame(name=c("\xC9lise","Pierre","Fran\xE7ois"),attendance=c(8,10,9))
+  res <- suppressWarnings( left_join(x, y, by = "name") )
+  expect_equal( nrow(res), 3L)
+  expect_equal( res$x, y$x)
+  
+  x <- data_frame(name=c("\xC9lise","Pierre","Fran\xE7ois"),score=c(5,7,6))
+  y <- data_frame(name=factor(c("\xC9lise","Pierre","Fran\xE7ois")),attendance=c(8,10,9))
+  res <- suppressWarnings( left_join(x, y, by = "name") )
+  expect_equal( nrow(res), 3L)
+  expect_equal( res$x, x$x)
+})
+